loader: efi comparison between signed and unsigned integer expressions
../../../common/multiboot2.c:273:18: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
   if (fp->f_size != st.st_size) {
                  ^~
../memmap.c:176:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i < smaplen; i++)
                ^
../main.c:670:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i < sz / sizeof (boot_order[0]); i++)
                ^
../main.c: In function 'command_chain':
../main.c:1121:21: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     if (pd->pd_unit == d_dev->d_slice) {
                     ^~
../copy.c:80:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i < count; i++, map = NextMemoryDescriptor(map, dsize)) {
                ^
../libefi.c:43:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i < ST->NumberOfTableEntries; i++) {
                ^
../env.c:535:17: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
   for (i = 0; i < datasz / sizeof (UINT16); i++)
                 ^
../efipart.c:144:19: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
   if (pd->pd_unit == dev->d_unit)
                   ^~
../efipart.c: In function 'efipart_readwrite':
../efipart.c:995:21: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (blk < 0 || blk > blkio->Media->LastBlock)
                     ^
../efipart.c:997:24: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if ((blk + nblks - 1) > blkio->Media->LastBlock)
                        ^
../efinet.c:159:40: error: signed and unsigned type in conditional expression [-Werror=sign-compare]
  return ((status == EFI_SUCCESS) ? len : -1);
                                        ^
../efinet.c: In function 'efinet_dev_init':
../efinet.c:313:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i < sz / sizeof (EFI_HANDLE); i++) {
                ^
../efi_console.c:216:34: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
   if (mode != -1 && console_mode != mode)
                                  ^~
../efi_console.c: In function 'efi_text_cons_display':
../efi_console.c:337:14: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (row - 1 == da->row && da->col == 0 && da->width == col)
              ^~
../efi_console.c:337:54: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (row - 1 == da->row && da->col == 0 && da->width == col)
                                                      ^~
../efi_console.c: In function 'efi_cons_init':
../efi_console.c:494:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i <= conout->Mode->MaxMode; i++) {
                ^~

+++ usr/src/boot/sys/boot/efi/loader/main.c
@@ -701,7 +701,7 @@ main(int argc, CHAR16 *argv[])
 	sz = sizeof (boot_order);
 	efi_global_getenv("BootOrder", &boot_order, &sz);
 	printf("   BootOrder:");
-	for (i = 0; i < sz / sizeof (boot_order[0]); i++)
+	for (i = 0; i < (int)(sz / sizeof (boot_order[0])); i++)
 		printf(" %04x%s", boot_order[i],
 		    boot_order[i] == boot_current ? "[*]" : "");
 	printf("\n");
@@ -1152,7 +1152,7 @@ command_chain(int argc, char *argv[])
 				/*
 				 * d_partition should be 255
 				 */
-				if (pd->pd_unit == d_dev->d_slice) {
+				if (pd->pd_unit == (uint32_t)d_dev->d_slice) {
 					loaded_image->DeviceHandle =
 					    pd->pd_handle;
 					break;
